removed pop up and added working make a request
